+++ bookkeeping/RELEASES.md
@@ -101,3 +101,13 @@ Based on:
 - [csharp v4.0.0] bookkeeping
 ### Releases
 - [NuGet v4.0.0] https://www.nuget.org/packages/WingspanBookkeeping/4.0.0 - bookkeeping
+
+## 2024-02-12 01:26:43
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.177.1 (2.252.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v4.1.0] bookkeeping
+### Releases
+- [NuGet v4.1.0] https://www.nuget.org/packages/WingspanBookkeeping/4.1.0 - bookkeeping